requirements.txt
flask
flask-cors
python-dotenv
requests

openai
langchain
langchain-community
langchain-core
langchain-text-splitters

pypdf
docx2txt
faiss-cpu
pandas
nltk
gensim
